feat: update list component
