+++ src/index.ts
@@ -29,7 +29,7 @@ export * from './component/view/Carousel'
 export * from './component/view/HorizontalScroller'
 export * from './component/view/Image'
 export * from './component/view/Input'
-// export * from './component/view/List'
+export * from './component/view/List'
 export * from './component/view/Scroller'
 export * from './component/view/Switch'
 export * from './component/view/Text'
